+++ docs/conf.py
@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-#
 # django-formtools documentation build configuration file, created by
 # sphinx-quickstart on Mon Oct  6 21:51:30 2014.
 #
@@ -46,8 +44,8 @@ source_suffix = '.rst'
 master_doc = 'index'
 
 # General information about the project.
-project = u'django-formtools'
-copyright = u'2014, Django Software Foundation and individual contributors'
+project = 'django-formtools'
+copyright = '2014, Django Software Foundation and individual contributors'
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the
@@ -197,8 +195,8 @@ latex_elements = {
 # Grouping the document tree into LaTeX files. List of tuples
 # (source start file, target name, title, author, documentclass [howto/manual]).
 latex_documents = [
-  ('index', 'django-formtools.tex', u'django-formtools Documentation',
-   u'Django Software Foundation and individual contributors', 'manual'),
+  ('index', 'django-formtools.tex', 'django-formtools Documentation',
+   'Django Software Foundation and individual contributors', 'manual'),
 ]
 
 # The name of an image file (relative to this directory) to place at the top of
@@ -227,8 +225,8 @@ latex_documents = [
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
 man_pages = [
-    ('index', 'django-formtools', u'django-formtools Documentation',
-     [u'Django Software Foundation and individual contributors'], 1)
+    ('index', 'django-formtools', 'django-formtools Documentation',
+     ['Django Software Foundation and individual contributors'], 1)
 ]
 
 # If true, show URL addresses after external links.
@@ -241,8 +239,8 @@ man_pages = [
 # (source start file, target name, title, author,
 #  dir menu entry, description, category)
 texinfo_documents = [
-  ('index', 'django-formtools', u'django-formtools Documentation',
-   u'Django Software Foundation and individual contributors', 'django-formtools', 'One line description of project.',
+  ('index', 'django-formtools', 'django-formtools Documentation',
+   'Django Software Foundation and individual contributors', 'django-formtools', 'One line description of project.',
    'Miscellaneous'),
 ]
 
@@ -262,10 +260,10 @@ texinfo_documents = [
 # -- Options for Epub output ---------------------------------------------------
 
 # Bibliographic Dublin Core info.
-epub_title = u'django-formtools'
-epub_author = u'Django Software Foundation and individual contributors'
-epub_publisher = u'Django Software Foundation and individual contributors'
-epub_copyright = u'2014, Django Software Foundation and individual contributors'
+epub_title = 'django-formtools'
+epub_author = 'Django Software Foundation and individual contributors'
+epub_publisher = 'Django Software Foundation and individual contributors'
+epub_copyright = '2014, Django Software Foundation and individual contributors'
 
 # The language of the text. It defaults to the language option
 # or en if the language is not set.